s2g: refactor how jobs & paths are handled
This changes no existing behavior, but is in preparation for adding
support for a new kind of more flexible jobfile syntax.

The changes in handle_path()/v1_handle_path() look big but are mostly
indentation because it now takes a specific 'job' (aka "operation")
instead of a list of jobs/operations.