handle SVG viewport units and viewBox
This commit makes svg2gcode deal correctly with SVGs containing units
on the viewport ("height" and "width" attributes), and SVGs with viewBox
scaling of the user units.

Specifically, this commit adds support when parsing the viewport for
all the units supported by the SVG spec:

    * px (96 dpi)
    * pt (72 dpi)
    * pc (6 dpi)
    * in
    * cm
    * mm

Also adds parsing of the viewBox and setting the X and Y scales, and
updates the "to_mm()" functions to work right.