add a new test: short-path-with-ramp
This test highlights a problem when using a ramp that's so shallow that
we reach the end of the path before the end of the ramp: the user's
requested depth of cut is not honored.

This is especially problematic when helix-milling holes with a diameter
only slightly larger than the cutter diameter: It ends up using a very
shallow depth of cut.