s2g: add --gcode-origin
This controls the relationship between the G-code coordinate system
origin and the origin of the SVG coordinate system.

Before this commit, gcoder always put the G-code origin at the lower
left corner of the SVG viewbox.  This makes "intuitive" sense, since it
makes the "artistic" gcode appear in the upper right quadrant instead
of the lower right.

But it is problematic when you use multiple SVGs that are all supposed
to overlay and have their features match up.

After this commit, the user can choose to make the SVG *origin* coincide
with the gcode origin, on the assumption that the origin is a point we
can all agree on.